fix(circular-dependency): Extract listener config and remove handler resolver from replies listener
