+++ starters/async-commons-starter/src/main/java/org/reactivecommons/async/starter/listeners/NotificationEventsListenerConfig.java
@@ -6,11 +6,12 @@ import org.reactivecommons.async.starter.config.ConnectionManager;
 import org.reactivecommons.async.starter.broker.BrokerProvider;
 import org.reactivecommons.async.starter.config.DomainHandlers;
 import org.reactivecommons.async.starter.config.ReactiveCommonsConfig;
+import org.reactivecommons.async.starter.config.ReactiveCommonsListenersConfig;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 @Configuration
-@Import(ReactiveCommonsConfig.class)
+@Import({ReactiveCommonsConfig.class, ReactiveCommonsListenersConfig.class})
 public class NotificationEventsListenerConfig extends AbstractListenerConfig {
 
     public NotificationEventsListenerConfig(ConnectionManager manager, DomainHandlers handlers) {
