+++ starters/async-commons-starter/src/main/java/org/reactivecommons/async/starter/senders/DirectAsyncGatewayConfig.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.java.Log;
 import org.reactivecommons.async.api.DirectAsyncGateway;
 import org.reactivecommons.async.starter.config.ConnectionManager;
-import org.reactivecommons.async.starter.config.DomainHandlers;
 import org.reactivecommons.async.starter.config.ReactiveCommonsConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,10 +19,10 @@ import java.util.concurrent.ConcurrentMap;
 public class DirectAsyncGatewayConfig {
 
     @Bean
-    public DirectAsyncGateway genericDirectAsyncGateway(ConnectionManager manager, DomainHandlers handlers) {
+    public DirectAsyncGateway genericDirectAsyncGateway(ConnectionManager manager) {
         ConcurrentMap<String, DirectAsyncGateway> directAsyncGateways = new ConcurrentHashMap<>();
         manager.forDomain((domain, provider) -> directAsyncGateways.put(domain,
-                provider.getDirectAsyncGateway(handlers.get(domain))));
+                provider.getDirectAsyncGateway()));
         return new GenericDirectAsyncGateway(directAsyncGateways);
     }
 }
